requirements.txt
#python3.7
psycopg2-binary
